AtomIDNet/requirements.txt
ase==3.22.0
numpy==1.21.1
opencv_python>4.6.0.0
pandas==1.3.1
Pillow==9.3.0
PyQt5==5.15.7
scikit_image==0.19.2
tensorboardX==2.5.1
torch==1.9.1+cu111
torchvision==0.10.1+cu111
tqdm==4.64.0
